Separate rendering from generation
We can generate a reverse colorblind test image independently from
rendering it to the screen. In this case we are rendering it to a
JavaFX canvas, but we could use other JavaFX UI components or even
other technologies like Swing as well, without affecting the code
responsible for generating the reverse colorblind test image.